Workaround to avoid nethcti-server crash at start
When nethcti-server is started on NS8, it get report api endpoint from
report configuration file. But when it tries to login to it, it fails
and crashes. Using the localhost url (http://127.0.0.1:$REPORTS_API_PORT)
instead of traefik url (https://$VHOST.$FQDN/pbx-report-api') works, but
the host is reachable from the container.